+++ src/components/Container/@interfaces/index.ts
@@ -1,0 +1,5 @@
+import { ReactNode } from 'react';
+
+export interface IContainerProps {
+  children: ReactNode;
+}
