+++ src/components/Container/index.tsx
@@ -1,0 +1,9 @@
+import { IContainerProps } from './@interfaces';
+
+export default function Container({ children }: IContainerProps) {
+  return (
+    <div className="flex flex-col justify-center items-center">
+      {children}
+    </div>
+  );
+}
