+++ src/components/Platform/@interfaces/index.ts
@@ -1,0 +1,7 @@
+import { MovieType } from '../../../@types/pages/Home/Movies';
+
+export interface IPlatformProps {
+  movies: MovieType[];
+  movieIndex: number;
+  handleMovie: () => void;
+}
