+++ src/components/Title/@interfaces/index.ts
@@ -1,0 +1,6 @@
+import { MovieType } from '../../../@types/pages/Home/Movies';
+
+export interface ITitleProps {
+  movies: MovieType[];
+  movieIndex: number;
+}
