+++ src/pages/_app.tsx
@@ -13,7 +13,9 @@ function MyApp({ Component, pageProps }: AppProps) {
         <meta name="description" content="Randovie is a random movie generator" />
       </Head>
 
-      <Component {...pageProps} />
+      <main className="bg-gray-800 h-screen w-screen flex flex-col justify-center items-center">
+        <Component {...pageProps} />
+      </main>
     </>
   );
 }
